Added welcome banner in catalog and renamed __version__ by meta
