+++ services/catalog/src/simcore_service_catalog/__init__.py
@@ -1,4 +1,4 @@
 """ Python package for the simcore_service_catalog.
 
 """
-from .__version__ import __version__
+from .meta import __version__
