+++ services/catalog/src/simcore_service_catalog/core/application.py
@@ -4,9 +4,9 @@ from typing import Callable, Optional
 
 from fastapi import FastAPI, Request
 
-from ..__version__ import api_version, api_vtag
 from ..api.root import router as api_router
 from ..api.routes.health import router as health_router
+from ..meta import api_version, api_vtag
 from .events import create_start_app_handler, create_stop_app_handler
 from .settings import AppSettings, BootModeEnum
 
@@ -43,6 +43,10 @@ def init_app(settings: Optional[AppSettings] = None) -> FastAPI:
 
     app.add_event_handler("startup", create_start_app_handler(app))
     app.add_event_handler("shutdown", create_stop_app_handler(app))
+
+    # setup app --
+    app.add_event_handler("startup", on_startup)
+    app.add_event_handler("shutdown", on_shutdown)
 
     # app.add_exception_handler(HTTPException, http_error_handler)
     # app.add_exception_handler(RequestValidationError, http422_error_handler)
