+++ services/catalog/src/simcore_service_catalog/core/events.py
@@ -5,12 +5,36 @@ from typing import Callable
 from fastapi import FastAPI
 
 from ..db.events import close_db_connection, connect_to_db
+from ..meta import __version__, project_name
 from ..services.director import close_director, setup_director
 from ..services.remote_debug import setup_remote_debugging
 from .background_tasks import start_registry_sync_task, stop_registry_sync_task
 from .settings import BootModeEnum
 
 logger = logging.getLogger(__name__)
+
+#
+# SEE https://patorjk.com/software/taag/#p=display&h=0&f=Ogre&t=Catalog
+#
+WELCOME_MSG = r"""
+   ___         _           _
+  / __\  __ _ | |_   __ _ | |  ___    __ _
+ / /    / _` || __| / _` || | / _ \  / _` |
+/ /___ | (_| || |_ | (_| || || (_) || (_| |
+\____/  \__,_| \__| \__,_||_| \___/  \__, |
+                                     |___/     {0}
+""".format(
+    f"v{__version__}"
+)
+
+
+def on_startup() -> None:
+    print(WELCOME_MSG, flush=True)
+
+
+def on_shutdown() -> None:
+    msg = project_name + f" v{__version__} SHUT DOWN"
+    print(f"{msg:=^100}", flush=True)
 
 
 def create_start_app_handler(app: FastAPI) -> Callable:
