+++ services/catalog/src/simcore_service_catalog/meta.py
@@ -19,8 +19,4 @@ try:
 except FileNotFoundError:
     metadata = current_distribution.get_metadata_lines("PKG-INFO")
 
-summary: str = next(
-    x.split(":")
-    for x in metadata
-    if x.startswith("Summary:")
-)[-1]
+summary: str = next(x.split(":") for x in metadata if x.startswith("Summary:"))[-1]
